Move javascript into separate file
